refactor: rename package to edu.agh.roomie and restructure application files
