--- backend/src/main/kotlin/Routing.kt
@@ -1,33 +1,0 @@
-package edu.agh
-
-import io.ktor.client.*
-import io.ktor.client.engine.apache.*
-import io.ktor.http.*
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.application.*
-import io.ktor.server.auth.*
-import io.ktor.server.plugins.compression.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.plugins.defaultheaders.*
-import io.ktor.server.plugins.openapi.*
-import io.ktor.server.plugins.statuspages.*
-import io.ktor.server.plugins.swagger.*
-import io.ktor.server.request.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-import io.ktor.server.sessions.*
-import org.jetbrains.exposed.sql.*
-
-fun Application.configureRouting() {
-  install(StatusPages) {
-    exception<Throwable> { call, cause ->
-      call.respondText(text = "500: $cause", status = HttpStatusCode.InternalServerError)
-    }
-  }
-  routing {
-    get("/") {
-      call.respondText("Hello World!")
-    }
-  }
-}
