+++ backend/src/main/kotlin/edu/agh/roomie/Application.kt
@@ -1,5 +1,9 @@
-package edu.agh
+package edu.agh.roomie
 
+import edu.agh.edu.agh.roomie.configureHTTP
+import edu.agh.edu.agh.roomie.configureRouting
+import edu.agh.edu.agh.roomie.configureSecurity
+import edu.agh.edu.agh.roomie.configureSerialization
 import io.ktor.server.application.*
 
 fun main(args: Array<String>) {
