+++ backend/src/main/kotlin/edu/agh/roomie/Databases.kt
@@ -1,4 +1,4 @@
-package edu.agh
+package edu.agh.roomie
 
 import io.ktor.client.*
 import io.ktor.client.engine.apache.*
