+++ backend/src/main/kotlin/edu/agh/roomie/Routing.kt
@@ -1,0 +1,20 @@
+package edu.agh.edu.agh.roomie
+
+import io.ktor.http.*
+import io.ktor.server.application.*
+import io.ktor.server.plugins.statuspages.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+
+fun Application.configureRouting() {
+  install(StatusPages) {
+    exception<Throwable> { call, cause ->
+      call.respondText(text = "500: $cause", status = HttpStatusCode.InternalServerError)
+    }
+  }
+  routing {
+    get("/") {
+      call.respondText("Hello World!")
+    }
+  }
+}
