+++ backend/src/main/kotlin/edu/agh/roomie/Security.kt
@@ -1,23 +1,13 @@
-package edu.agh
+package edu.agh.edu.agh.roomie
 
 import io.ktor.client.*
 import io.ktor.client.engine.apache.*
 import io.ktor.http.*
-import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
-import io.ktor.server.plugins.compression.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.plugins.defaultheaders.*
-import io.ktor.server.plugins.openapi.*
-import io.ktor.server.plugins.statuspages.*
-import io.ktor.server.plugins.swagger.*
-import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import io.ktor.server.sessions.*
-import org.jetbrains.exposed.sql.*
 
 fun Application.configureSecurity() {
   authentication {
