+++ backend/src/main/kotlin/edu/agh/roomie/Serialization.kt
@@ -1,0 +1,13 @@
+package edu.agh.edu.agh.roomie
+
+import io.ktor.server.application.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+
+fun Application.configureSerialization() {
+  routing {
+    get("/json/kotlinx-serialization") {
+      call.respond(mapOf("hello" to "world"))
+    }
+  }
+}
