+++ backend/src/main/kotlin/edu/agh/roomie/UserScheme.kt
@@ -1,4 +1,4 @@
-package edu.agh
+package edu.agh.roomie
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.serialization.Serializable
